pyproject.toml
[project]
name = "anabasis-assistent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "baml-py>=0.219.0",
    "discord-py>=2.7.1",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.5",
    "mypy>=1.19.1",
    "pandas>=3.0.1",
    "psycopg>=3.3.3",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]


[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
mypy_path = ["src"]
explicit_package_bases = true
